Revised Expand to Area
- Simplified and corrected PolygonEx.ExpandToArea
- Moved PolygonEx.ExpandToArea test to PolygonExTests
